requirements.txt
# Sentinel DDoS — Python Dependencies

# ── Core ─────────────────────────────────────────
fastapi==0.110.3
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Storage ──────────────────────────────────────
redis[hiredis]==5.0.4
sqlalchemy[asyncio]==2.0.30
aiosqlite==0.20.0

# ── AI / ML ──────────────────────────────────────
scikit-learn==1.4.2
numpy==1.26.4

# ── Rules ────────────────────────────────────────
pyyaml==6.0.1

# ── Utilities ────────────────────────────────────
python-dotenv==1.0.1
python-multipart==0.0.9
